requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
SQLAlchemy>=2.0.30,<3.0
python-multipart==0.0.9
faster-whisper==1.0.2
google-generativeai==0.7.2
httpx==0.27.0
python-dotenv==1.0.1
